Address PR review feedback
- Forward the original request to ASSETS.fetch in the markdown middleware
  so conditional caching (If-None-Match -> 304) keeps working, and treat
  304 as success instead of falling back to HTML
- Keep site-relative image paths (e.g. /company-logos/...) as site URLs
  instead of routing them through Cloudflare Images
- Build technology-page ADR sections from project ADR cards, which carry
  their origin project, instead of de-duplicating on bare ADR slugs that
  are only unique within a project
- Extract shared killProcessGroup/waitForServer wrangler test helpers and
  add the early-exit handler to the markdown negotiation test
