+++ ui/tests/integration/posthog-proxy.test.ts
@@ -1,6 +1,7 @@
 import { type ChildProcess, spawn } from "node:child_process";
 import { createServer, type Server } from "node:http";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
+import { killProcessGroup, waitForServer } from "./wrangler-test-utils";
 
 const WRANGLER_PORT = 8788;
 const MOCK_API_PORT = 8789;
@@ -175,38 +176,3 @@ describe("PostHog Proxy Integration Test", () => {
     expect(forwardedRequest?.url).toBe("/static/array.js");
   });
 });
-
-function killProcessGroup(proc: ChildProcess | undefined): void {
-  if (!proc?.pid) return;
-  try {
-    process.kill(-proc.pid, "SIGTERM");
-  } catch {
-    proc.kill();
-  }
-}
-
-async function waitForServer(
-  proc: ChildProcess,
-  timeout: number,
-): Promise<void> {
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      killProcessGroup(proc);
-      reject(new Error(`Server did not start within ${timeout}ms`));
-    }, timeout);
-
-    proc.stdout?.on("data", (data: Buffer) => {
-      if (data.toString().includes("Ready on")) {
-        clearTimeout(timer);
-        resolve();
-      }
-    });
-
-    proc.on("exit", (code) => {
-      clearTimeout(timer);
-      if (code !== 0 && code !== null) {
-        reject(new Error(`Wrangler process exited with code ${code}`));
-      }
-    });
-  });
-}
